requirements.txt
flask
line-bot-sdk
anthropic
requests
python-dotenv
gunicorn
